Refactor metrics merging to work with endpoints-controller
Previously, all metrics configuration was dealt with in the mutating
webhook handler. Now, since service/proxy registration happens in
endpoints-controller, some of the metrics configuration needs to be used
in endpoints-controller as well. There is still some functionality in
the handler that also needs metrics configuration, such as adding
prometheus annotations and running the consul sidecar.

This refactor pulls out common configuration for metrics into a
MetricsConfig struct and adds the methods for getting values from
flags and annotations to that struct, so it can be commonly used between
endpoints-controller and the webhook handler.
