+++ connect-inject/consul_sidecar.go
@@ -1,7 +1,6 @@
 package connectinject
 
 import (
-	"errors"
 	"fmt"
 
 	corev1 "k8s.io/api/core/v1"
@@ -12,36 +11,19 @@ import (
 // It always disables service registration because for connect we no longer
 // need to keep services registered as this is handled in the endpoints-controller.
 func (h *Handler) consulSidecar(pod corev1.Pod) (corev1.Container, error) {
-	run, err := h.shouldRunMergedMetricsServer(pod)
+	metricsPorts, err := h.MetricsConfig.mergedMetricsServerConfiguration(pod)
 	if err != nil {
 		return corev1.Container{}, err
 	}
-
-	// This should never happen because we only call this function in the handler if
-	// we need to run the metrics merging server. This check is here just in case.
-	if !run {
-		return corev1.Container{}, errors.New("metrics merging should be enabled in order to inject the consul-sidecar")
-	}
-
-	// Configure consul sidecar with the appropriate metrics flags.
-	mergedMetricsPort, err := h.mergedMetricsPort(pod)
-	if err != nil {
-		return corev1.Container{}, err
-	}
-	serviceMetricsPath := h.serviceMetricsPath(pod)
-
-	// Don't need to check the error since it's checked in the call to
-	// h.shouldRunMergedMetricsServer() above.
-	serviceMetricsPort, _ := h.serviceMetricsPort(pod)
 
 	command := []string{
 		"consul-k8s",
 		"consul-sidecar",
 		"-enable-service-registration=false",
 		"-enable-metrics-merging=true",
-		fmt.Sprintf("-merged-metrics-port=%s", mergedMetricsPort),
-		fmt.Sprintf("-service-metrics-port=%s", serviceMetricsPort),
-		fmt.Sprintf("-service-metrics-path=%s", serviceMetricsPath),
+		fmt.Sprintf("-merged-metrics-port=%s", metricsPorts.mergedPort),
+		fmt.Sprintf("-service-metrics-port=%s", metricsPorts.servicePort),
+		fmt.Sprintf("-service-metrics-path=%s", metricsPorts.servicePath),
 	}
 
 	return corev1.Container{
