+++ connect-inject/consul_sidecar_test.go
@@ -6,35 +6,19 @@ import (
 	"github.com/hashicorp/go-hclog"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-var (
-	consulSidecarResources = corev1.ResourceRequirements{
-		Requests: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse("10m"),
-			corev1.ResourceMemory: resource.MustParse("25Mi"),
-		},
-		Limits: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse("20m"),
-			corev1.ResourceMemory: resource.MustParse("50Mi"),
-		},
-	}
-)
-
-// NOTE: This is tested here rather than in handler_test because doing it there
-// would require a lot of boilerplate to get at the underlying patches that would
-// complicate understanding the tests (which are simple).
 
 // Test that if the conditions for running a merged metrics server are true,
 // that we pass the metrics flags to consul sidecar.
 func TestConsulSidecar_MetricsFlags(t *testing.T) {
 	handler := Handler{
-		Log:                         hclog.Default().Named("handler"),
-		ImageConsulK8S:              "hashicorp/consul-k8s:9.9.9",
-		DefaultEnableMetrics:        true,
-		DefaultEnableMetricsMerging: true,
+		Log:            hclog.Default().Named("handler"),
+		ImageConsulK8S: "hashicorp/consul-k8s:9.9.9",
+		MetricsConfig: MetricsConfig{
+			DefaultEnableMetrics:        true,
+			DefaultEnableMetricsMerging: true,
+		},
 	}
 	container, err := handler.consulSidecar(corev1.Pod{
 		ObjectMeta: metav1.ObjectMeta{
@@ -58,23 +42,4 @@ func TestConsulSidecar_MetricsFlags(t *testing.T) {
 	require.Contains(t, container.Command, "-merged-metrics-port=20100")
 	require.Contains(t, container.Command, "-service-metrics-port=8080")
 	require.Contains(t, container.Command, "-service-metrics-path=/metrics")
-}
-
-// Test that the Consul sidecar errors when metrics merging is disabled.
-func TestConsulSidecar_ErrorsWhenMetricsMergingIsDisabled(t *testing.T) {
-	handler := Handler{
-		Log:                    hclog.Default().Named("handler"),
-		ImageConsulK8S:         "hashicorp/consul-k8s:9.9.9",
-		ConsulSidecarResources: consulSidecarResources,
-	}
-	_, err := handler.consulSidecar(corev1.Pod{
-		Spec: corev1.PodSpec{
-			Containers: []corev1.Container{
-				{
-					Name: "web",
-				},
-			},
-		},
-	})
-	require.EqualError(t, err, "metrics merging should be enabled in order to inject the consul-sidecar")
 }
