+++ connect-inject/handler.go
@@ -22,10 +22,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
-const (
-	defaultServiceMetricsPath = "/metrics"
-)
-
 var (
 	codecs       = serializer.NewCodecFactory(runtime.NewScheme())
 	deserializer = codecs.UniversalDeserializer()
@@ -113,14 +109,10 @@ type Handler struct {
 	DefaultProxyMemoryRequest resource.Quantity
 	DefaultProxyMemoryLimit   resource.Quantity
 
-	// Default metrics settings. These will configure where Prometheus scrapes
-	// metrics from, and whether to run a merged metrics endpoint on the consul
-	// sidecar. These can be overridden via pod annotations.
-	DefaultEnableMetrics        bool
-	DefaultEnableMetricsMerging bool
-	DefaultMergedMetricsPort    string
-	DefaultPrometheusScrapePort string
-	DefaultPrometheusScrapePath string
+	// MetricsConfig contains metrics configuration from the inject-connect command and has methods to determine whether
+	// configuration should come from the default flags or annotations. The handler uses this to configure prometheus
+	// annotations and the merged metrics server.
+	MetricsConfig MetricsConfig
 
 	// Resource settings for init container. All of these fields
 	// will be populated by the defaults provided in the initial flags.
@@ -218,7 +210,7 @@ func (h *Handler) Handle(_ context.Context, req admission.Request) admission.Res
 	// (that functionality lives in the endpoints-controller),
 	// we only need the consul sidecar to run the metrics merging server.
 	// First, determine if we need to run the metrics merging server.
-	shouldRunMetricsMerging, err := h.shouldRunMergedMetricsServer(pod)
+	shouldRunMetricsMerging, err := h.MetricsConfig.shouldRunMergedMetricsServer(pod)
 	if err != nil {
 		h.Log.Error("Error determining if metrics merging server should be run", "err", err, "Request Name", req.Name)
 		return admission.Errored(http.StatusInternalServerError, fmt.Errorf("error determining if metrics merging server should be run: %s", err))
@@ -352,101 +344,18 @@ func (h *Handler) defaultAnnotations(pod *corev1.Pod) error {
 	return nil
 }
 
-// enableMetrics returns the default value in the handler, or overrides that
-// with the annotation if provided.
-func (h *Handler) enableMetrics(pod corev1.Pod) (bool, error) {
-	enabled := h.DefaultEnableMetrics
-	if raw, ok := pod.Annotations[annotationEnableMetrics]; ok && raw != "" {
-		enableMetrics, err := strconv.ParseBool(raw)
-		if err != nil {
-			return false, fmt.Errorf("%s annotation value of %s was invalid: %s", annotationEnableMetrics, raw, err)
-		}
-		enabled = enableMetrics
-	}
-	return enabled, nil
-}
-
-// enableMetricsMerging returns the default value in the handler, or overrides
-// that with the annotation if provided.
-func (h *Handler) enableMetricsMerging(pod corev1.Pod) (bool, error) {
-	enabled := h.DefaultEnableMetricsMerging
-	if raw, ok := pod.Annotations[annotationEnableMetricsMerging]; ok && raw != "" {
-		enableMetricsMerging, err := strconv.ParseBool(raw)
-		if err != nil {
-			return false, fmt.Errorf("%s annotation value of %s was invalid: %s", annotationEnableMetricsMerging, raw, err)
-		}
-		enabled = enableMetricsMerging
-	}
-	return enabled, nil
-}
-
-// mergedMetricsPort returns the default value in the handler, or overrides
-// that with the annotation if provided.
-func (h *Handler) mergedMetricsPort(pod corev1.Pod) (string, error) {
-	return determineAndValidatePort(pod, annotationMergedMetricsPort, h.DefaultMergedMetricsPort, false)
-}
-
-// prometheusScrapePort returns the default value in the handler, or overrides
-// that with the annotation if provided.
-func (h *Handler) prometheusScrapePort(pod corev1.Pod) (string, error) {
-	return determineAndValidatePort(pod, annotationPrometheusScrapePort, h.DefaultPrometheusScrapePort, false)
-}
-
-// prometheusScrapePath returns the default value in the handler, or overrides
-// that with the annotation if provided.
-func (h *Handler) prometheusScrapePath(pod corev1.Pod) string {
-	if raw, ok := pod.Annotations[annotationPrometheusScrapePath]; ok && raw != "" {
-		return raw
-	}
-
-	return h.DefaultPrometheusScrapePath
-}
-
-// serviceMetricsPort returns the port the service exposes metrics on. This will
-// default to the port used to register the service with Consul, and can be
-// overridden with the annotation if provided.
-func (h *Handler) serviceMetricsPort(pod corev1.Pod) (string, error) {
-	// The annotationPort is the port used to register the service with Consul.
-	// If that has been set, it'll be used as the port for getting service
-	// metrics as well, unless overridden by the service-metrics-port annotation.
-	if raw, ok := pod.Annotations[annotationPort]; ok && raw != "" {
-		// The service metrics port can be privileged if the service author has
-		// written their service in such a way that it expects to be able to use
-		// privileged ports. So, the port metrics are exposed on the service can
-		// be privileged.
-		return determineAndValidatePort(pod, annotationServiceMetricsPort, raw, true)
-	}
-
-	// If the annotationPort is not set, the serviceMetrics port will be 0
-	// unless overridden by the service-metrics-port annotation. If the service
-	// metrics port is 0, the consul sidecar will not run a merged metrics
-	// server.
-	return determineAndValidatePort(pod, annotationServiceMetricsPort, "0", true)
-}
-
-// serviceMetricsPath returns a default of /metrics, or overrides
-// that with the annotation if provided.
-func (h *Handler) serviceMetricsPath(pod corev1.Pod) string {
-	if raw, ok := pod.Annotations[annotationServiceMetricsPath]; ok && raw != "" {
-		return raw
-	}
-
-	return defaultServiceMetricsPath
-}
-
-// prometheusAnnotations returns the Prometheus scraping configuration
-// annotations. It returns a nil map if metrics are not enabled and annotations
-// should not be set.
+// prometheusAnnotations sets the Prometheus scraping configuration
+// annotations on the Pod.
 func (h *Handler) prometheusAnnotations(pod *corev1.Pod) error {
-	enableMetrics, err := h.enableMetrics(*pod)
+	enableMetrics, err := h.MetricsConfig.enableMetrics(*pod)
 	if err != nil {
 		return err
 	}
-	prometheusScrapePort, err := h.prometheusScrapePort(*pod)
+	prometheusScrapePort, err := h.MetricsConfig.prometheusScrapePort(*pod)
 	if err != nil {
 		return err
 	}
-	prometheusScrapePath := h.prometheusScrapePath(*pod)
+	prometheusScrapePath := h.MetricsConfig.prometheusScrapePath(*pod)
 
 	if enableMetrics {
 		pod.Annotations[annotationPrometheusScrape] = "true"
@@ -454,68 +363,6 @@ func (h *Handler) prometheusAnnotations(pod *corev1.Pod) error {
 		pod.Annotations[annotationPrometheusPath] = prometheusScrapePath
 	}
 	return nil
-}
-
-// shouldRunMergedMetricsServer returns whether we need to run a merged metrics
-// server. This is used to configure the consul sidecar command, and the init
-// container, so it can pass appropriate arguments to the consul connect envoy
-// command.
-func (h *Handler) shouldRunMergedMetricsServer(pod corev1.Pod) (bool, error) {
-	enableMetrics, err := h.enableMetrics(pod)
-	if err != nil {
-		return false, err
-	}
-	enableMetricsMerging, err := h.enableMetricsMerging(pod)
-	if err != nil {
-		return false, err
-	}
-	serviceMetricsPort, err := h.serviceMetricsPort(pod)
-	if err != nil {
-		return false, err
-	}
-
-	// Don't need to check error here since serviceMetricsPort has been
-	// validated by calling h.serviceMetricsPort above
-	smp, _ := strconv.Atoi(serviceMetricsPort)
-
-	if enableMetrics && enableMetricsMerging && smp > 0 {
-		return true, nil
-	}
-	return false, nil
-}
-
-// determineAndValidatePort behaves as follows:
-// If the annotation exists, validate the port and return it.
-// If the annotation does not exist, return the default port.
-// If the privileged flag is true, it will allow the port to be in the
-// privileged port range of 1-1023. Otherwise, it will only allow ports in the
-// unprivileged range of 1024-65535.
-func determineAndValidatePort(pod corev1.Pod, annotation string, defaultPort string, privileged bool) (string, error) {
-	if raw, ok := pod.Annotations[annotation]; ok && raw != "" {
-		port, err := portValue(pod, raw)
-		if err != nil {
-			return "", fmt.Errorf("%s annotation value of %s is not a valid integer", annotation, raw)
-		}
-
-		if privileged && (port < 1 || port > 65535) {
-			return "", fmt.Errorf("%s annotation value of %d is not in the valid port range 1-65535", annotation, port)
-		} else if !privileged && (port < 1024 || port > 65535) {
-			return "", fmt.Errorf("%s annotation value of %d is not in the unprivileged port range 1024-65535", annotation, port)
-		}
-
-		// if the annotation exists, return the validated port
-		return fmt.Sprint(port), nil
-	}
-
-	// if the annotation does not exist, return the default
-	if defaultPort != "" {
-		port, err := portValue(pod, defaultPort)
-		if err != nil {
-			return "", fmt.Errorf("%s is not a valid port on the pod %s", defaultPort, pod.Name)
-		}
-		return fmt.Sprint(port), nil
-	}
-	return "", nil
 }
 
 // consulNamespace returns the namespace that a service should be
